fix: make XML cache invalidation content-aware
