--- .github/workflows/fix-deep-bug-sweep-ci.yml
@@ -1,47 +1,0 @@
-name: Fix deep bug sweep CI
-# synchronize trigger: workflow already existed in the previous PR head
-
-on:
-  pull_request:
-    branches: [ "master" ]
-    paths: [ ".github/workflows/fix-deep-bug-sweep-ci.yml" ]
-
-permissions:
-  contents: write
-
-jobs:
-  fix:
-    if: github.event.pull_request.head.repo.full_name == github.repository
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1
-        with:
-          ref: fix/deep-bug-sweep
-          fetch-depth: 0
-      - name: Repair Config scope and clean temporary files
-        shell: bash
-        run: |
-          set -euo pipefail
-          python3 - <<'PY'
-          from pathlib import Path
-
-          path = Path('service/src/main/java/cleveres/tricky/cleverestech/Config.kt')
-          text = path.read_text()
-          old = '''        }
-    val cacheIterator = storedKeyboxCache.keys.iterator()
-'''
-          new = '''        }
-    }
-    val cacheIterator = storedKeyboxCache.keys.iterator()
-'''
-          count = text.count(old)
-          if count != 1:
-              raise SystemExit(f'expected one Config scope match, found {count}')
-          path.write_text(text.replace(old, new, 1))
-          PY
-          rm -f ci-trigger.txt ci-diagnostics.txt .github/workflows/fix-deep-bug-sweep-ci.yml
-          git config user.name "github-actions[bot]"
-          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-          git add -A
-          git commit -m "fix: repair keybox cache scope"
-          git push origin HEAD:fix/deep-bug-sweep
