--- ci-trigger.txt
@@ -1,1 +1,0 @@
-trigger one-shot source repair
